refactor: reorganize files and split code - split api.ts into chunks so that we have better segregation of code - new folder structure
