+++ src/collection-manager/collection-manager.test.ts
@@ -3,10 +3,10 @@ import { objectToCamelCase, objectToSnakeCase } from "@thinknimble/tn-utils"
 import axios from "axios"
 import { beforeEach, describe, expect, it, Mocked, vi } from "vitest"
 import { z } from "zod"
-import { Pagination } from "./pagination"
-import { createApi } from "./api"
-import { createCollectionManager } from "./collection-manager"
-import { getPaginatedSnakeCasedZod } from "./utils/pagination"
+import { Pagination } from "../utils/pagination"
+import { createApi } from "../api"
+import { createCollectionManager } from "."
+import { getPaginatedSnakeCasedZod } from "../utils/pagination"
 
 vi.mock("axios")
